# Disable bashlog.sh unless TFENV_DEBUG is higher than 0