Bashlog puts an unnecessary overhead on all terraform commands.
If debug is not enabled, don't use bashlog, use a more lightweight
log function instead.

Fixes #196 and #229