@reegnz 2021-12-14
### Benchmarks
#### Before
```sh
❯ time ./terraform -version
Terraform v1.1.0
on darwin_amd64
./terraform -version  0.44s user 0.49s system 62% cpu 1.492 total
```
#### After
```sh
❯ time ./terraform -version
Terraform v1.1.0
on darwin_amd64
./terraform -version  0.06s user 0.04s system 77% cpu 0.130 total
```

I also made sure to keep the color coding of the log levels.