@reegnz 2021-12-14
I also prepared #313 which is probably better than this one, as that is definitely backward-compatible (so no major version bump needed).